requirements.txt
asgiref==3.8.1
Django==5.0.3
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
